+++ src/plugin.cpp
@@ -119,6 +119,7 @@ void init(Plugin* p) {
   p->addModel(modelThru);
   p->addModel(modelVCAMix4);
   p->addModel(modelVCAMix4Stereo);
+  p->addModel(modelVCOUnit);
   p->addModel(modelVenomBlank);
   p->addModel(modelWidgetMenuExtender);
   p->addModel(modelWinComp);